Blackjack casino game using HTML CSS and JavaScript
This is a classis Blackjack game in which player will be assigned to card at the start of the game and in-order to win the game the user had to have a sum of 21.
If initially he had the sum of 21 i.e. he has an ACE and any of the face card then he had won the game. If less then 21 then he have to draw a another card using the new card button.
The new card button will assign another card to the user and it will automatically be added to  user's sum. After drawing the card if user have the sum of 21 then he had won the game and if he had more than 21 then he is out of the game, and if less than 21 then the game continues.